Support different conformance level testing
Old behaviour: ics was a single integer string that listed the
maximum conformance level and all levels below that defined level
are also tested.

New behaviour: ics is a string of conformance levels joined by
commas, listing each individual conformance level to check. This
is the behaviour defined in the HTML documentation for ETS.