Fix failing tests and embedded sample CDB
Sample CDB had incorrect Component Selectors in the filenames of its
fake files.
Some tests were applying CS validation to only a single dataset instead
of multiple.
One of the geocells in the sample CDB was removed to reduce the
maintenance work in keeping its fake files up to date.