Fix "LC" LODs being incorrectly failed in tests
Old behaviour was to assume all LODs less than 0 are a) in the "LC"
directory, and b) have the LOD "LC" in their filename. This is
incorrect according to the spec (v1.0 and v1.1).

Instead, coarse LODs are in the "LC" directory *and* the filenames
follow the "LC\d\d" pattern. This commit fixes the sample data,
JUnit tests, and TestNG methods to properly evaluate these
filenames.