Embed sample CDB resources in artifact
This ensures that the sample CDB (and its reference XML files) is
always included in the build artifacts, and is available for usage
by different platforms/environments. Test classes will be updated
to refer to this sample CDB.

The resource declarations from ets-common are here as there is no
simple merging of these elements with the parent POM file.