Normalize IUT path under TEAM Engine
When ran under TEAM Engine, the IUT *may* include `file:///` at the
start of the string. This causes issues as different test methods
in the ETS are not expecting that string. By removing this string
if present, tests now pass correctly when ran under TEAM Engine as
well as Eclipse.

This fixes #26.